# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Małopolska Social Innovation Hub — Matchmaking API",
    "version": "1.0.0",
    "description": "AI-assisted matchmaking between a described social problem and existing social innovations, documented cases and contextual evidence. Relevance means suitability for the described need, not a probability of success."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/api/v1/tests": {
      "get": {
        "tags": [
          "Tester"
        ],
        "operationId": "listInnovationTests",
        "summary": "List innovation tests open to testers",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TestStatus"
            }
          },
          {
            "name": "innovationId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 10000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of innovation tests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tester"
        ],
        "operationId": "createInnovationTest",
        "summary": "Create an innovation test (operator)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Test created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "data"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/InnovationTest"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tests/{testId}": {
      "get": {
        "tags": [
          "Tester"
        ],
        "operationId": "getInnovationTest",
        "summary": "Get test details, application count and remaining slots",
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/InnovationTest"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Test not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/problem-reports": {
      "get": {
        "tags": [
          "Problem reports"
        ],
        "operationId": "listProblemReports",
        "summary": "List reported social problems with their review status",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ProblemReportStatus"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 10000,
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of problem reports.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemReportListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Problem reports"
        ],
        "operationId": "createProblemReport",
        "summary": "Report a local social problem (residents, NGOs, institutions)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemReportRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Problem report accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "data"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProblemReport"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/problem-reports/{reportId}": {
      "get": {
        "tags": [
          "Problem reports"
        ],
        "operationId": "getProblemReport",
        "summary": "Get a single problem report and its review status",
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Problem report.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProblemReport"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Report not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Problem reports"
        ],
        "operationId": "updateProblemReport",
        "summary": "Update the review status / response (administrator)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProblemReportUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report updated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message",
                    "data"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ProblemReport"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Administrator role required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Report not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TestStatus": {
        "type": "string",
        "enum": [
          "recruiting",
          "active",
          "completed",
          "cancelled"
        ]
      },
...
      "InnovationTest": {
        "type": "object",
        "required": [
          "id",
          "innovationId",
          "title",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "innovationId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "instructions": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "maxTesters": {
            "type": [
              "integer",
              "null"
            ]
          },
          "startAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "endAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/TestStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "innovationTitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "innovationSummary": {
            "type": [
              "string",
              "null"
            ]
          },
          "applicationsCount": {
            "type": "integer",
            "minimum": 0
          },
          "acceptedCount": {
            "type": "integer",
            "minimum": 0
          },
          "slotsLeft": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          }
        }
      },
      "TestListResponse": {
        "type": "object",
        "required": [
          "total",
          "count",
          "limit",
          "offset",
          "data"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InnovationTest"
            }
          }
        }
      },
      "CreateTestRequest": {
        "type": "object",
        "required": [
          "innovationId",
          "title"
        ],
        "additionalProperties": false,
        "properties": {
          "innovationId": {
            "type": "string",
            "maxLength": 200
          },
          "title": {
            "type": "string",
            "maxLength": 300
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "instructions": {
            "type": "string",
            "maxLength": 5000
          },
          "location": {
            "type": "string",
            "maxLength": 300
          },
          "maxTesters": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/TestStatus"
          }
        }
      },
...
      "ProblemReportStatus": {
        "type": "string",
        "enum": [
          "new",
          "in_review",
          "planned",
          "resolved",
          "rejected"
        ]
      },
      "ProblemReportReporter": {
        "type": "string",
        "enum": [
          "resident",
          "ngo",
          "local_government",
          "institution",
          "other"
        ]
      },
      "ProblemReport": {
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "reporterType",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "municipality": {
            "type": [
              "string",
              "null"
            ]
          },
          "county": {
            "type": [
              "string",
              "null"
            ]
          },
          "reporterType": {
            "$ref": "#/components/schemas/ProblemReportReporter"
          },
          "status": {
            "$ref": "#/components/schemas/ProblemReportStatus"
          },
          "adminResponse": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProblemReportRequest": {
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "minLength": 3,
            "maxLength": 300
          },
          "description": {
            "type": "string",
            "minLength": 20,
            "maxLength": 5000
          },
          "category": {
            "type": "string",
            "maxLength": 160
          },
          "municipality": {
            "type": "string",
            "maxLength": 160
          },
          "county": {
            "type": "string",
            "maxLength": 160
          },
          "reporterType": {
            "$ref": "#/components/schemas/ProblemReportReporter"
          },
          "contactEmail": {
            "type": "string",
            "format": "email",
            "maxLength": 320
          }
        }
      },
      "ProblemReportUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ProblemReportStatus"
          },
          "adminResponse": {
            "type": "string",
            "maxLength": 5000
          }
        }
      },
      "ProblemReportListResponse": {
        "type": "object",
        "required": [
          "total",
          "count",
          "limit",
          "offset",
          "data"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "minimum": 1
          },
          "offset": {
            "type": "integer",
            "minimum": 0
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemReport"
            }
          }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId",
              "retryable"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string",
                "description": "Safe Polish message."
              },
              "requestId": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}